pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Personal Agentic RAG system — Docker edition"
authors = [
    { name = "Shreyas1015", email = "shreyas1234gurav@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery[redis]>=5.6.3",
    "fastapi[standard]>=0.136.1",
    "fastembed>=0.8.0",
    "langchain>=1.2.15",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.9",
    "llama-index>=0.14.21",
    "llama-index-readers-docling>=0.4.2",
    "logto>=0.2.1",
    "openai>=2.32.0",
    "pyjwt[crypto]>=2.12.1",
    "qdrant-client>=1.17.1",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.49",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]
